# apis/docs/swagger.yaml
definitions:
...
  domain.ForgotRequest:
    properties:
      confirm_pass:
        type: string
      email:
        type: string
      new_pass:
        type: string
    required:
    - confirm_pass
    - email
    - new_pass
    type: object
...
  domain.Response:
    properties:
      message:
        type: string
    type: object
...
  domain.TwoPhaseRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
...
info:
  contact: {}
paths:
...
  /forgot:
    put:
      consumes:
      - application/json
      description: Esqueci minha senha
      parameters:
      - description: Dados da comunidade
        in: body
        name: Corpo
        required: true
        schema:
          $ref: '#/definitions/domain.ForgotRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Sucesso
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Informações inválidas
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Esqueci minha senha
      tags:
      - Esqueci minha senha
  /forgot/verify:
    post:
      consumes:
      - application/json
      description: Esqueci minha senha
      parameters:
      - description: Dados da verificação
        in: body
        name: Corpo
        required: true
        schema:
          $ref: '#/definitions/domain.TwoPhaseRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Sucesso
          schema:
            type: ""
        "400":
          description: Informações inválidas
          schema:
            $ref: '#/definitions/domain.Response'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Realiza a verificação de duplo fator do esqueci minha senha
      tags:
      - Esqueci minha senha
...
swagger: "2.0"